# OAS/openapi.yaml
openapi: "3.0.0"
servers:
  - url: /ori-mock
info:
  title: "De ORI API voor Open Overheid"
  description: "Gegevens uit de Open Raads Informatie systemen kunnen met deze API worden aangeleverd voor actieve openbaarmaking via Open Overheid. Ook kan deze API gebruikt worden om aangeleverde data bij te werken. Deze versie van deze API zal naast documentobjecten ook vergader-informatie kunnen verwerken volgens het ontwerp van de VNG mini-ORI API, en daarbij ook een endpoint voor Vergaderingen en Agendapunten ontsluiten. Bij gebruik door decentrale overheden zal de DigiKoppeling-standaard verplicht worden voor het aanleveren bij Open Overheid (zie Forum Standaardisatie). Nu wordt nog de Client Credentials flow toegepast. ([OAuth 2.0](https://swagger.io/docs/specification/authentication/oauth2/)) Deze specificatie wordt met een uitgebreide toelichting onderhouden op https://gitlab.com/koop/woo/aanleveren-ori "
  version: "0.1.0"
  x-imvertor-generator-version: "2.0.0"
  x-yamlCompiler-stylesheets-version: "20230315"
  contact:
    url: "https://gitlab.com/koop/woo/aanleveren-ori"
  license:
    name: "European Union Public License, version 1.2 (EUPL-1.2)"
    url: "https://eupl.eu/1.2/nl/"
paths:
  /agendapunten:
    get:
      operationId: "getagendapunten"
      summary: "Alle agendapunten opvragen."
      description: "Deze lijst kan in de toekomst gefilterd worden met behulp van query parameters."
      responses:
        200:
          description: "Zoekactie geslaagd"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAgendapuntList"
        401:
          description: "Je bent niet geautoriseerd voor deze aanvraag."
        404:
          description: "De gevraagde resource is niet gevonden."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      tags:
        - "Agendapunten"
    post:
      operationId: "postagendapunt"
      description: "Het vastleggen van een agendapunt"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgendapuntZonderPid"
      responses:
        201:
          description: "Succesvol aangeleverd."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agendapunt"
        400:
          description: "Aanleverfout"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: "Je bent niet geautoriseerd voor deze aanvraag."
      tags:
        - "Agendapunten"
  /agendapunten/{id}:
    delete:
      operationId: "delagendapunt"
      description: "Het bericht voor het verwijderen van een agendapunt."
      parameters:
        - in: path
          name: id
          schema:
            type: string
            description: "Unieke resource identifier."
          required: true
      responses:
        200:
          description: "Verwijderactie geslaagd"
        401:
          description: "Je bent niet geautoriseerd voor deze aanvraag."
        404:
          description: "De gevraagde resource is niet gevonden."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      tags:
        - "Agendapunten"
    get:
      operationId: "getagendapunt"
      summary: "Een specifiek agendapunt opvragen."
      description: "Een specifiek agendapunt opvragen op basis van een identifier."
      parameters:
        - in: path
          name: id
          schema:
            type: string
            description: "Unieke resource identifier."
          required: true
      responses:
        200:
          description: "Zoekactie geslaagd"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agendapunt"
        401:
          description: "Je bent niet geautoriseerd voor deze aanvraag."
        404:
          description: "De gevraagde resource is niet gevonden."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      tags:
        - "Agendapunten"
    put:
      operationId: "putagendapunt"
      description: "Het wijzigen van een agendapunt"
      parameters:
        - in: path
          name: id
          schema:
            type: string
            description: "Unieke resource identifier."
          required: true
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - properties:
                    organisatie:
                      readOnly: true
                - $ref: "#/components/schemas/Agendapunt"
      responses:
        201:
          description: "Succesvol gewijzigd."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agendapunt"
        400:
          description: "Aanleverfout"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: "Je bent niet geautoriseerd voor deze aanvraag."
        404:
          description: "De gevraagde resource is niet gevonden."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        409:
          description: "Deze wijziging bevat aanpassingen op velden die niet mogen worden aangepast."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      tags:
        - "Agendapunten"
...
  /vergaderingen/{id}:
    delete:
      operationId: "delvergadering"
      description: "Het bericht voor het verwijderen van een vergadering"
      parameters:
        - in: path
          name: id
          schema:
            type: string
            description: "Unieke resource identifier."
            title: ID
          required: true
      responses:
        200:
          description: "Verwijderactie geslaagd"
        401:
          description: "Je bent niet geautoriseerd voor deze aanvraag."
        404:
          description: "De gevraagde resource is niet gevonden."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      tags:
        - "Vergaderingen"
    get:
      operationId: "getvergadering"
      summary: "Een specifieke vergadering opvragen."
      description: "Een specifieke vergadering opvragen op basis van een identifier."
      parameters:
        - in: path
          name: id
          schema:
            type: string
            description: "Unieke resource identifier."
            title: ID
          required: true
      responses:
        200:
          description: "Zoekactie geslaagd"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vergadering"
        401:
          description: "Je bent niet geautoriseerd voor deze aanvraag."
        404:
          description: "De gevraagde resource is niet gevonden."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      tags:
        - "Vergaderingen"
    put:
      operationId: "putvergadering"
      description: "Het wijzigen van een vergadering"
      parameters:
        - in: path
          name: id
          schema:
            type: string
            description: "Unieke resource identifier."
          required: true
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - properties:
                    organisatie:
                      readOnly: true
                - $ref: "#/components/schemas/Vergadering"
      responses:
        201:
          description: "Succesvol gewijzigd."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vergadering"
        400:
          description: "Aanleverfout"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: "Je bent niet geautoriseerd voor deze aanvraag."
        404:
          description: "De gevraagde resource is niet gevonden."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        409:
          description: "Deze wijziging bevat aanpassingen op velden die niet mogen worden aangepast."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      tags:
        - "Vergaderingen"
components:
  schemas:
    Agendapunt:
      type: "object"
      description: "Een onderdeel van de agenda dat als één geheel wordt behandeld. Een agendapunt kan weer onderverdeeld zijn naar subagendapunten."
      required:
        - "pid"
        - "organisatie"
        - "dossiertype"
        - "agendapuntnaam"
        - "vergadering"
      properties:
        pid:
          type: string
          format: uri
          readOnly: true
          title: url
          description: "Persistent Identifier, URL-referentie naar dit object binnen Open Overheid. Dit is de unieke identificatie en locatie van dit object."
          minLength: 1
        webpaginalink:
          title: Webpagina link
          description: "De URL naar de webpagina waar het agendapunt is terug te vinden."
          type: "string"
          format: "uri"
          example: "https://example.com/mijn-pagina.html"
        organisatie:
          $ref: "#/components/schemas/Organisatie"
        dossiertype:
          title: Dossiertype
          description: "Het type dossier, hier te vullen met 'agendapunt'."
          type: "string"
          example: "agendapunt"
        agendapuntnaam:
          type: "string"
          title: "Agendapunt Naam"
          description: "De weergave van het Agendapunt op de agenda. De letterlijke volgorde van een Agendapunt kan ook worden aangevuld met attribuut 'Volgnummer' of 'tussenKop'. Hoe het Agendapunt getoond wordt op de Agenda wordt beschreven met dit attribuut."
        vergadering:
          allOf:
            - description: "Verwijzing naar de vergadering waar dit agendapunt behandeld wordt."
            - $ref: "#/components/schemas/VerwijzingNaarResource"
        hoofdagendapunt:
          allOf:
            - description: "Verwijzing naar het hoofdagendapunt waarvan dit het subagendapunt is."
            - $ref: "#/components/schemas/VerwijzingNaarResource"
        omschrijving:
          type: "string"
          title: "Agendapunt omschrijving"
          description: "De omschrijving van het agendapunt"
        volgnummer:
          type: "string"
          title: "agendapuntvolgnummer"
          description: "Dit is het volgnummer van het agendapunt tijdens de vergadering. Dit is de letterlijke volgorde en zegt niets over hoe dit volgnummer getoond wordt. Hoe het Agendapunt getoond wordt op de Agenda wordt beschreven met attribuut 'agendapuntnaam'."
        subagendapunten:
          type: array
          items:
            type: string
            format: uri
          description: "URL-referenties naar de subagendapunten."
          title: Subagendapunten
          uniqueItems: true
        tussenkop:
          type: "string"
          title: "Tussenkop"
          description: "Eventuele andere sorteringsmechanismen om de agendapunten te rangschikken."
        overig:
          type: "string"
          title: "overig"
          description: "Vrij veld, bijvoorbeeld voor het noemen van status of om andere informatie toe te voegen."
        starttijd:
          type: "string"
          title: "Starttijd"
          description: "De datum en tijdstip waarop de behandeling van het agendapunt is gestart"
          format: "date-time"
          example: "2017-02-10T18:25:00Z"
        eindtijd:
          type: "string"
          title: "Eindtijd"
          description: "De datum en tijdstip waarop de behandeling van het agendapunt is geëindigd. Moet voorkomen als Starttijd voorkomt."
          format: "date-time"
          example: "2017-02-10T18:25:00Z"
        locatie:
          type: "string"
          title: "Locatie"
          description: "Omschrijving van de locatie waar het agendapunt wordt behandeld"
        geplandvolgnummer:
          type: "string"
          title: "Gepland Volgnummer"
          description: "Dit is het geplande volgnummer van het agendapunt van een vergadering."
        geplandeeindtijd:
          type: "string"
          title: "Geplande eindtijd"
          description: "De geplande datum en tijdstip waarop de behandeling van het agendapunt eindigt"
          format: "date-time"
          example: "2017-02-10T18:25:00Z"
        geplandestarttijd:
          type: "string"
          title: "Geplande starttijd"
          description: "De geplande datum en tijdstip waarop de behandeling van het agendapunt start"
          format: "date-time"
          example: "2017-02-10T18:25:00Z"
        indicatiehamerstuk:
          type: "boolean"
          title: "Indicatie hamerstuk"
          description: ""
          example: "true"
          nullable: true
        indicatiebehandeld:
          type: "boolean"
          title: "Indicatie behandeld"
          description: "Indicatie geeft aan of het agendapunt is behandeld tijdens de vergadering. Is eventueel af te leiden als Starttijd en/of Eindtijd zijn ingevuld."
          nullable: true
        indicatiebesloten:
          type: "boolean"
          title: "Indicatie besloten"
          description: Indicatie of een agendapunt besloten wordt behandeld.
          nullable: true
    AgendapuntZonderPid:
      type: "object"
      description: "Een onderdeel van de agenda dat als één geheel wordt behandeld. Een agendapunt kan weer onderverdeeld zijn naar subagendapunten. Bij aanlevering krijgt een agendapunt een Pid van Open Overheid."
      required:
        - "organisatie"
        - "dossiertype"
        - "agendapuntnaam"
        - "vergadering"
      properties:
        webpaginalink:
          title: Webpagina link
          description: "De URL naar de webpagina waar het agendapunt is terug te vinden."
          type: "string"
          format: "uri"
          example: "https://example.com/mijn-pagina.html"
        organisatie:
          $ref: "#/components/schemas/Organisatie"
        dossiertype:
          title: Dossiertype
          description: "Het type dossier, hier te vullen met 'agendapunt'."
          type: "string"
          example: "agendapunt"
        agendapuntnaam:
          type: "string"
          title: "Agendapunt Naam"
          description: "De weergave van het Agendapunt op de agenda. De letterlijke volgorde van een Agendapunt kan ook worden aangevuld met attribuut 'volgnummer' of 'tussenkop'. Hoe het Agendapunt getoond wordt op de Agenda wordt beschreven met dit attribuut."
        vergadering:
          allOf:
            - description: "Verwijzing naar de vergadering waar dit agendapunt behandeld wordt."
            - $ref: "#/components/schemas/VerwijzingNaarResource"
        hoofdagendapunt:
          allOf:
            - description: "Verwijzing naar het hoofdagendapunt waarvan dit het subagendapunt is."
            - $ref: "#/components/schemas/VerwijzingNaarResource"
        omschrijving:
          type: "string"
          title: "Agendapunt omschrijving"
          description: "De omschrijving van het agendapunt"
        volgnummer:
          type: "string"
          title: "Agendapuntvolgnummer"
          description: "Dit is het volgnummer van het agendapunt tijdens de vergadering. Dit is de letterlijke volgorde en zegt niets over hoe dit volgnummer getoond wordt. Hoe het Agendapunt getoond wordt op de Agenda wordt beschreven met attribuut 'Agendapuntnaam'."
        subagendapunten:
          type: array
          items:
            type: string
            format: uri
          description: "URL-referenties naar de subagendapunten."
          title: Subagendapunten
          uniqueItems: true
        tussenkop:
          type: "string"
          title: "Tussenkop"
          description: "Eventuele andere sorteringsmechanismen om de agendapunten te rangschikken."
        overig:
          type: "string"
          title: "overig"
          description: "Vrij veld, bijvoorbeeld voor het noemen van status of om andere informatie toe te voegen."
        starttijd:
          type: "string"
          title: "Starttijd"
          description: "De datum en tijdstip waarop de behandeling van de agendapunt is gestart"
          format: "date-time"
          example: "2017-02-10T18:25:00Z"
        eindtijd:
          type: "string"
          title: "Eindtijd"
          description: "De datum en tijdstip waarop de behandeling van de agendapunt is geëindigd. Moet voorkomen als Starttijd voorkomt."
          format: "date-time"
          example: "2017-02-10T18:25:00Z"
        locatie:
          type: "string"
          title: "Locatie"
          description: "Omschrijving van de locatie waar het agendapunt wordt behandeld"
        geplandvolgnummer:
          type: "string"
          title: "Gepland Volgnummer"
          description: "Dit is het geplande volgnummer van het agendapunt van een vergadering."
        geplandeeindtijd:
          type: "string"
          title: "Geplande eindtijd"
          description: "De geplande datum en tijdstip waarop de behandeling van het agendapunt eindigt"
          format: "date-time"
          example: "2017-02-10T18:25:00Z"
        geplandestarttijd:
          type: "string"
          title: "Geplande starttijd"
          description: "De geplande datum en tijdstip waarop de behandeling van het agendapunt start"
          format: "date-time"
          example: "2017-02-10T18:25:00Z"
        indicatiehamerstuk:
          type: "boolean"
          title: "Indicatie hamerstuk"
          description: ""
          example: "true"
          nullable: true
        indicatiebehandeld:
          type: "boolean"
          title: "Indicatie behandeld"
          description: "Indicatie geeft aan of het agendapunt is behandeld tijdens de vergadering. Is eventueel af te leiden als Starttijd en/of Eindtijd zijn ingevuld."
          nullable: true
        indicatiebesloten:
          type: "boolean"
          title: "Indicatie besloten"
          description: Indicatie of een agendapunt besloten wordt behandeld.
          nullable: true
...
    Vergadering:
      type: "object"
      description: "Een vergadering is een bijeenkomst van meerdere mensen die met elkaar spreken en/of afspraken maken over onderwerpen die op de agenda staan."
      required:
        - "pid"
        - "organisatie"
        - "dossiertype"
        - "naam"
      properties:
        pid:
          type: string
          format: uri
          readOnly: true
          title: url
          description: "Persistent Identifier, URL-referentie naar dit object binnen Open Overheid. Dit is de unieke identificatie en locatie van dit object."
          minLength: 1
        webpaginalink:
          title: Webpagina link
          description: "De URL naar de webpagina waar het agendapunt is terug te vinden."
          type: "string"
          format: "uri"
          example: "https://example.com/mijn-pagina.html"
        organisatie:
          $ref: "#/components/schemas/Organisatie"
        dossiertype:
          title: Dossiertype
          description: "Het type dossier, hier te vullen met 'vergadering'."
          type: "string"
          example: "vergadering"
        naam:
          type: "string"
          title: "Naam"
          description: "De naam van de vergadering"
          example: "Raadsvergadering"
        aanvang:
          type: "string"
          title: "Aanvang vergadering"
          description: "De datum en tijdstip waarop de vergadering is gestart. Komt zeker voor als Einde vergadering ook is ingevuld"
          format: "date-time"
          example: "2017-02-10T18:25:00Z"
        hoofdvergadering:
          allOf:
            - description: "Verwijzing naar de hoofdvergadering waarvan deze vergadering een deelvergadering is."
            - $ref: "#/components/schemas/VerwijzingNaarResource"
        einde:
          type: "string"
          title: "Einde"
          description: "De datum en tijdstip waarop de vergadering is geëindigd"
          format: "date-time"
          example: "2017-02-10T21:12:00Z"
        georganiseerddoorgremium:
          allOf:
            - $ref: "#/components/schemas/Gremium"
            - title: "georganiseerddoorgremium"
              description: "de vergadering is georganiseerd door een gremium"
        geplandeaanvang:
          type: "string"
          title: "Geplande aanvang"
          description: "De geplande datum en tijdstip waarop de vergadering start"
          format: "date-time"
          example: "2017-02-09T18:25:00Z"
        geplandeeinde:
          type: "string"
          title: "Geplande einde"
          description: "De geplande datum en tijdstip waarop de vergadering eindigt"
          format: "date-time"
          example: "2017-02-09T21:25:00Z"
        geplandedatum:
          type: "string"
          title: "Geplande vergaderdatum"
          description: "De geplande datum van de VERGADERING"
          format: "date"
          example: "2017-02-09"
        locatie:
          type: "string"
          title: "Locatie"
          description: "Omschrijving van de locatie waar de vergadering wordt gehouden"
          example: "Raadszaal, Stadskantoor, Overtoom 2"
        vergaderstatus:
          $ref: "#/components/schemas/VergaderingStatus"
        vergadertoelichting:
          type: "string"
          title: "Vergadertoelichting"
          description: "De toelichting of nadere omschrijving van de vergadering."
          example: "Regulier geplande raadsvergadering van mei"
        vergaderdatum:
          type: "string"
          title: "Vergaderdatum"
          description: "De datum van de Vergadering"
          format: "date"
          example: "2017-02-10"
        vergaderingstype:
          $ref: "#/components/schemas/VergaderingsType"
        deelvergaderingen:
          type: array
          items:
            type: string
            format: uri
            minLength: 1
            nullable: true
          description: "URL-referenties naar de deelvergaderingen van deze vergadering."
          title: Deelvergaderingen
          uniqueItems: true
...
    Gremium:
      type: "object"
      description: "De referentielijst met alle gremia in een gemeente. Hier horen ook alle commissies bij van de gemeenteraad."
      required:
        - "gremiumidentificatie"
        - "gremiumnaam"
      properties:
        gremiumidentificatie:
          type: "string"
          title: "Gremiumidentificatie"
          description: "De identificatie van een gremium"
          minLength: 1
        gremiumnaam:
          type: "string"
          title: "Gremiumnaam"
          description: "De naam van het gremium. Hier horen ook de (raads- en staten)commissies bij."
          minLength: 1
...
    VergaderingStatus:
      type: "string"
      description: "mogelijke status van een vergadering"
    VergaderingsType:
      type: "string"
      description: "de mogelijke typen van een vergadering"
    VerwijzingNaarResource:
      type: object
      required:
        - url
        - id
      properties:
        id:
          description: "ID van de resource waarnaar verwezen wordt."
          type: string
          minLength: 1
        url:
          description: "URL-referentie naar de resource waarnaar verwezen wordt."
          type: string
          format: uri
          minLength: 1
          readOnly: true
    Gemeente:
      type: object
      properties:
        gemeente:
          type: string
          title: Gemeente
          description: 'De code van de gemeente volgens TOOI.
            Bijvoorbeeld "gm0363" voor gemeente Amsterdam.'
          example: gm0363
        naam:
          type: string
          description: "De naam van de organisatie."
          example: "Gemeente Amsterdam"
      required:
        - gemeente
        - naam
    Provincie:
      type: object
      properties:
        provincie:
          type: string
          title: Provincie
          description: 'De code van de provincie volgens TOOI.
            Bijvoorbeeld "pv27" voor provincie Noord-Holland.'
          example: pv27
        naam:
          type: string
          description: "De naam van de organisatie."
          example: "Provincie Groningen"
      required:
        - provincie
        - naam
    Waterschap:
      type: object
      properties:
        waterschap:
          type: string
          title: Waterschap
          description: 'De code van het waterschap volgens TOOI.
            Bijvoorbeeld "ws0654" voor het waterschap Aa en Maas.'
          example: ws0654
        naam:
          type: string
          description: "De naam van de organisatie."
          example: "Waterschap Aa en Maas"
      required:
        - waterschap
        - naam
    Organisatie:
      oneOf:
        - $ref: "#/components/schemas/Gemeente"
        - $ref: "#/components/schemas/Provincie"
        - $ref: "#/components/schemas/Waterschap"
    PaginatedAgendapuntList:
      type: object
      properties:
        next:
          type: string
          format: uri
          nullable: true
          description: "URL naar de volgende pagina met resultaten."
        previous:
          type: string
          format: uri
          nullable: true
          description: "URL naar de vorige pagina met resultaten."
        results:
          type: array
          items:
            $ref: "#/components/schemas/Agendapunt"
...
    ErrorResponse:
      type: object
      properties:
        titel:
          type: string
          description: "Titel van de foutmelding."
        status:
          type: integer
          format: int32
          description: "De HTTP statuscode van het antwoord."
        detail:
          type: string
          description: "Omschrijving van de foutmelding."